feat: Replaced generated code with stellar-xdr
Removed the dependencies xdr-rs-serialize and xdr-rs-serialize-derive,
since serialization is already provided by stellar-xdr.

For compatibility reasons and convenience, the existing XDRSerialize and
XDRDeserialize traits are now implemented on top of the WriteXdr and
ReadXdr traits from stellar-xdr.

Added a bunch of new data types and enum values. Removed enum types
where they added nothing beyond the same type in stellar-xdr. No support
was added for Preconditions::V2 and TransactionExt::V1, both part of the
new stellar_xdr_curr::Transaction type. When trying to convert
transactions which use these values an UnsupportedFeature error is
returned for now.

Also, not all new types have been wrapped. For example, the inner data
of the new LedgerKey values instead refers directly to the types from
the stellar-xdr crate. This can be changed when there is demand, but it
seems futile to wrap all these types in verbose API and to keep this
updated.

Closes #10

Signed-off-by: Thorbjørn Lindeijer <bjorn@lindeijer.nl>